chore(trie): rename cursor module
